Fix slotid incorrect reading
SlotID is an int not a byte

+++ PvPModifier/Network/Packets/PlayerSlotArgs.cs
@@ -3,20 +3,24 @@ using System;
 using System.IO;
 using System.IO.Streams;
 
-namespace PvPModifier.Network.Packets {
+namespace PvPModifier.Network.Packets
+{
 
-    public class PlayerSlotArgs : EventArgs {
+    public class PlayerSlotArgs : EventArgs
+    {
         public PvPPlayer Player;
         public byte SlotId;
         public short Stack;
         public byte Prefix;
         public short NetID;
 
-        public bool ExtractData(MemoryStream data, PvPPlayer player, out PlayerSlotArgs arg) {
+        public bool ExtractData(MemoryStream data, PvPPlayer player, out PlayerSlotArgs arg)
+        {
             data.ReadByte(); //Passes through the PlayerID data
 
-            arg = new PlayerSlotArgs {
-                SlotId = (byte)data.ReadByte(),
+            arg = new PlayerSlotArgs
+            {
+                SlotId = (byte)data.ReadInt16(),
                 Player = player,
                 Stack = data.ReadInt16(),
                 Prefix = (byte)data.ReadByte(),
